chore: First pass at creating a download module
