+++ projects/Mallard/src/push-notifications/push-notifications.ts
@@ -1,13 +1,6 @@
 import moment, { MomentInput } from 'moment'
 import { Platform } from 'react-native'
 import PushNotification from 'react-native-push-notification'
-import {
-    clearOldIssues,
-    downloadAndUnzipIssue,
-    matchSummmaryToKey,
-} from 'src/helpers/files'
-import { imageForScreenSize } from 'src/helpers/screen'
-import { getIssueSummary } from 'src/hooks/use-issue-summary'
 import { pushNotificationRegistrationCache } from '../helpers/storage'
 import PushNotificationIOS from '@react-native-community/push-notification-ios'
 import { defaultSettings } from 'src/helpers/settings/defaults'
@@ -17,6 +10,7 @@ import ApolloClient from 'apollo-client'
 import { Feature } from 'src/services/logging'
 import { registerWithNotificationService } from './notification-service'
 import { notificationTracking } from './notification-tracking'
+import { downloadViaNotification } from 'src/download/downloadViaNotification'
 
 export interface PushNotificationRegistration {
     registrationDate: string
@@ -103,56 +97,11 @@ const pushNotifcationRegistration = (apolloClient: ApolloClient<object>) => {
 
             if (key) {
                 try {
-                    const screenSize = await imageForScreenSize()
-
-                    await pushTracking(
-                        'pushScreenSize',
-                        screenSize,
-                        Feature.DOWNLOAD,
-                    )
-
-                    const issueSummaries = await getIssueSummary()
-
-                    await pushTracking(
-                        'pushIssueSummaries',
-                        JSON.stringify(issueSummaries),
-                        Feature.DOWNLOAD,
-                    )
-
-                    // Check to see if we can find the image summary for the one that is pushed
-                    const pushImageSummary = matchSummmaryToKey(
-                        issueSummaries,
-                        key,
-                    )
-
-                    await pushTracking(
-                        'pushImageSummary',
-                        JSON.stringify(pushImageSummary),
-                        Feature.DOWNLOAD,
-                    )
-
-                    await downloadAndUnzipIssue(
-                        apolloClient,
-                        pushImageSummary,
-                        screenSize,
-                    )
-
-                    await pushTracking(
-                        'pushDownloadComplete',
-                        'completed',
-                        Feature.DOWNLOAD,
-                    )
+                    await downloadViaNotification(key, apolloClient)
                     notificationTracking(notificationId, 'downloaded')
                 } catch (e) {
-                    await pushTracking(
-                        'pushDownloadError',
-                        JSON.stringify(e),
-                        Feature.DOWNLOAD,
-                    )
                     errorService.captureException(e)
                 } finally {
-                    // No matter what happens, always clear up old issues
-                    await clearOldIssues()
                     // required on iOS only (see fetchCompletionHandler docs: https://facebook.github.io/react-native/docs/pushnotificationios.html)
                     notification.finish(PushNotificationIOS.FetchResult.NoData)
                 }
